+++ 01_主題課程/02_在 Discord 跟 AI 機器人聊天/practice/practice_2.py
@@ -14,6 +14,7 @@ while True:
     # 2. 允許使用者在終端機中輸入問題
     # 3. 允許使用者可以反覆輸入問題
     # 4. 允許使用者可以輸入 exit 或 quit 來結束對話
+    # 5. 請實作記憶功能
     # [Start] 請在以下空間撰寫機器人程式
